requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
spacy>=3.7.0
numpy>=1.24.0
pandas>=2.0.0

# Data processing
jsonlines>=3.1.0
tqdm>=4.65.0
pyyaml>=6.0.0

# Logging and monitoring
structlog>=23.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Prodigy integration
# prodigy>=1.12.0  # Uncomment if using Prodigy programmatically
